25/04/16

+++ sprweb3/src/main/java/pack/model/ScoreModel.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 public class ScoreModel {
 	public String score(ScoreBean bean) {
 		String data = bean.getName() + "님의 총점은 " + (bean.getJava()+bean.getPython()) + "점, 평균은 "
-					+ (bean.getJava()+bean.getPython())/2 + "점입니다." ;
+					+ (double)(bean.getJava()+bean.getPython())/2 + "점입니다." ;
 		return data;
 	}
 }
